src/requirements.txt
apache-airflow-providers-snowflake==5.8.0

